add type checks to tests
Unsurprisingly, this uncovered more tests that were actually failing to
return the right types. However, I'm actually more scared of the ones
that ARE passing -- these may well be suffering from the opposite
problem of failed type inference (-> any).
I realize this means it's becoming a lot harder to rely on CI for PRs
this way -- sorry about that. That said, I imagine it's better not to
leave these under the rug, even if it's checking only half the potential
problems (types too strict rather than too lenient). If I'm changing
things, I'd like to get a stronger sense of whether it's for better or
worse.