Use prepack instead of prepublish
`prepublish` is no longer the correct way to build before `publish`. The correct script to use is `prepack`. The reason for this is that `prepublish` runs after `pack` and before `publish`, and `pack` is the script which defines what is published during `publish`, i.e., the package being published is already created by the time `prepublish` is called. By building in `prepack` the package will be created properly during `pack` and then published during `publish`. Using `prepublish` relies on having an already-done, possibly stale build.